mempool/monitor.go
package mempool

import (
	"context"
	"errors"
	"fmt"
	"math/big"
	"github.com/michaelpento.lv/mevbot/config"
	"github.com/michaelpento.lv/mevbot/mempool/network"
	"github.com/michaelpento.lv/mevbot/mempool/network/dpdk"
	"runtime"
	"sync"
	"sync/atomic"
	"time"

	"github.com/ethereum/go-ethereum/common"
	"github.com/ethereum/go-ethereum/core/types"
	lru "github.com/hashicorp/golang-lru"
	"github.com/prometheus/client_golang/prometheus"
	"go.uber.org/zap"
	"golang.org/x/time/rate"
	"github.com/iovisor/gobpf/bcc"
	"github.com/iovisor/gobpf/bpf"
)

type Subscription interface {
	Err() <-chan error
	Unsubscribe()
}

type WorkerConfig struct {
	Workers   int
	QueueSize int
}

type MempoolMonitor struct {
	cfg    *config.Config
	client EthClient
	logger *zap.Logger

	// Enhanced transaction handling
	txChan    chan *Transaction
	txPool    *MMapPool
	txIndexer *MempoolIndexer

	// Performance optimization
	workers   *AffinityWorkerPool
	limiter   *rate.Limiter
	cache     *lru.Cache
	breaker   *CircuitBreaker
	ebpfMon   *ebpf.SyscallMonitor

	// Network optimization
	netManager *dpdk.DPDKManager

	// Metrics and monitoring
	metrics struct {
		transactionCount prometheus.CounterVec
		blockCount       prometheus.CounterVec
		txLatency        prometheus.Histogram
		profitTotal      prometheus.Counter
		profitPerTx      prometheus.Histogram
		memoryUsage      prometheus.Gauge
		goroutineCount   prometheus.Gauge
		opportunityCount prometheus.CounterVec
		errorCount       prometheus.Counter
		queueDepth       prometheus.Gauge
	}

	// Concurrency control
	mu     sync.RWMutex
	ctx    context.Context
	cancel context.CancelFunc
	wg     sync.WaitGroup
}

type EthClient interface {
	SubscribePendingTransactions(ctx context.Context, ch chan<- common.Hash) (Subscription, error)
	TransactionByHash(ctx context.Context, hash common.Hash) (*types.Transaction, bool, error)
	Close()
}

type Transaction struct {
	*types.Transaction
	FirstSeen time.Time
	GasPrice  *big.Int
	Priority  float64 // Priority score for processing
}

func NewMempoolMonitor(cfg *config.Config, client EthClient, logger *zap.Logger) (*MempoolMonitor, error) {
	ctx, cancel := context.WithCancel(context.Background())

	// Create memory-mapped transaction pool
	txPool, err := NewMMapPool("/tmp/mevbot_mempool.mmap")
	if err != nil {
		return nil, fmt.Errorf("failed to create mmap pool: %w", err)
	}

	// Create CPU-pinned worker pool
	workers := NewAffinityWorkerPool(&WorkerConfig{
		Workers:   runtime.NumCPU(),  // Use all available CPUs
		QueueSize: 10000,            // Large queue for burst handling
	})

	// Create eBPF monitor
	ebpfMon, err := ebpf.NewSyscallMonitor(logger)
	if err != nil {
		txPool.Close()
		return nil, fmt.Errorf("failed to create eBPF monitor: %w", err)
	}

	// Create DPDK network manager
	netManager, err := dpdk.NewDPDKManager(logger, &dpdk.Config{
		Interface:    "eth0",  // Main network interface
		NumRxQueues: 4,       // Use 4 RX queues
		NumTxQueues: 4,       // Use 4 TX queues
		PollTimeout: time.Microsecond * 100,
	})
	if err != nil {
		txPool.Close()
		ebpfMon.Stop()
		return nil, fmt.Errorf("failed to create DPDK manager: %w", err)
	}

	monitor := &MempoolMonitor{
		cfg:        cfg,
		client:     client,
		logger:     logger,
		ctx:        ctx,
		cancel:     cancel,
		txChan:     make(chan *Transaction, 10000),
		txPool:     txPool,
		workers:    workers,
		ebpfMon:    ebpfMon,
		netManager: netManager,
		limiter:    rate.NewLimiter(rate.Limit(cfg.RateLimit), cfg.RateBurst),
		breaker:    NewCircuitBreaker(cfg.CircuitBreaker, logger),
	}

	// Initialize metrics
	monitor.initMetrics()

	return monitor, nil
}

func (m *MempoolMonitor) Start(ctx context.Context) chan *Transaction {
	m.wg.Add(1)
	// Start CPU-pinned worker pool
	if err := m.workers.Start(); err != nil {
		m.logger.Error("Failed to start worker pool", zap.Error(err))
		return nil
	}

	// Start eBPF monitoring
	if err := m.ebpfMon.Start(ctx); err != nil {
		m.logger.Error("Failed to start eBPF monitor", zap.Error(err))
		m.workers.Stop()
		return nil
	}

	// Start DPDK network manager
	if err := m.netManager.Start(ctx); err != nil {
		m.logger.Error("Failed to start DPDK manager", zap.Error(err))
		m.workers.Stop()
		m.ebpfMon.Stop()
		return nil
	}

	go m.monitorLoop(ctx)
	go m.collectSystemMetrics(ctx)
	return m.txChan
}

func (m *MempoolMonitor) monitorLoop(ctx context.Context) {
	defer m.wg.Done()

	// Subscribe to pending transactions
	pendingTxHashes := make(chan common.Hash, 1024)
	sub, err := m.client.SubscribePendingTransactions(ctx, pendingTxHashes)
	if err != nil {
		m.logger.Error("Failed to subscribe to pending transactions", zap.Error(err))
		return
	}
	defer sub.Unsubscribe()

	// Process pending transactions
	for {
		select {
		case <-ctx.Done():
			return
		case err := <-sub.Err():
			m.logger.Error("Subscription error", zap.Error(err))
			m.breaker.RecordError(err)
			return
		case hash := <-pendingTxHashes:
			// Fetch full transaction
			tx, _, err := m.client.TransactionByHash(ctx, hash)
			if err != nil {
				m.logger.Error("Failed to get transaction", zap.Error(err))
				continue
			}

			// Process transaction
			transaction := &Transaction{
				Transaction: tx,
				FirstSeen:  time.Now(),
				GasPrice:   tx.GasPrice(),
				Priority:   calculatePriority(tx),
			}

			if err := m.processPendingTx(transaction); err != nil {
				m.logger.Error("Failed to process transaction", zap.Error(err))
			}
		}
	}
}

func (m *MempoolMonitor) processPendingTx(tx *Transaction) error {
	if !m.isRelevantTransaction(tx) {
		return nil
	}

	if !m.breaker.IsHealthy() {
		return errors.New("circuit breaker is tripped")
	}

	if err := m.limiter.Wait(m.ctx); err != nil {
		return fmt.Errorf("rate limiter error: %w", err)
	}

	// Record metrics
	m.metrics.transactionCount.WithLabelValues("processed").Inc()
	start := time.Now()
	defer func() {
		m.metrics.txLatency.Observe(float64(time.Since(start).Milliseconds()))
	}()

	// Process transaction
	if err := m.txIndexer.IndexTransaction(tx); err != nil {
		m.metrics.errorCount.Inc()
		return fmt.Errorf("failed to index transaction: %w", err)
	}

	return nil
}

func (m *MempoolMonitor) Shutdown() error {
	m.cancel()
	// Stop worker pool
	if err := m.workers.Stop(); err != nil {
		m.logger.Error("Failed to stop worker pool", zap.Error(err))
	}

	// Stop eBPF monitor
	if err := m.ebpfMon.Stop(); err != nil {
		m.logger.Error("Failed to stop eBPF monitor", zap.Error(err))
	}

	// Stop DPDK manager
	if err := m.netManager.Stop(); err != nil {
		m.logger.Error("Failed to stop DPDK manager", zap.Error(err))
	}

	// Close memory-mapped pool
	if err := m.txPool.Close(); err != nil {
		m.logger.Error("Failed to close mmap pool", zap.Error(err))
	}

	m.wg.Wait()
	close(m.txChan)
	return nil
}

func (m *MempoolMonitor) initMetrics() {
	m.metrics.transactionCount = *prometheus.NewCounterVec(prometheus.CounterOpts{
		Name: "mempool_transactions_total",
		Help: "Total number of transactions processed by type",
	}, []string{"type"})

	m.metrics.blockCount = *prometheus.NewCounterVec(prometheus.CounterOpts{
		Name: "mempool_blocks_total",
		Help: "Total number of blocks processed",
	}, []string{"type"})

	m.metrics.txLatency = prometheus.NewHistogram(prometheus.HistogramOpts{
		Name:    "mempool_transaction_latency_seconds",
		Help:    "Histogram of transaction processing latencies",
		Buckets: prometheus.ExponentialBuckets(0.001, 2, 10), // Start at 1ms, double 10 times
	})

	m.metrics.profitTotal = prometheus.NewCounter(prometheus.CounterOpts{
		Name: "mempool_profit_total_wei",
		Help: "Total profit in wei",
	})

	m.metrics.profitPerTx = prometheus.NewHistogram(prometheus.HistogramOpts{
		Name:    "mempool_profit_per_tx_wei",
		Help:    "Histogram of profit per transaction in wei",
		Buckets: prometheus.ExponentialBuckets(1e9, 10, 10), // Start at 1 Gwei, multiply by 10 each time
	})

	m.metrics.memoryUsage = prometheus.NewGauge(prometheus.GaugeOpts{
		Name: "mempool_memory_usage_bytes",
		Help: "Current memory usage in bytes",
	})

	m.metrics.goroutineCount = prometheus.NewGauge(prometheus.GaugeOpts{
		Name: "mempool_goroutine_count",
		Help: "Current number of goroutines",
	})

	m.metrics.opportunityCount = *prometheus.NewCounterVec(
		prometheus.CounterOpts{
			Name: "mempool_opportunities_total",
			Help: "Total number of opportunities by type",
		},
		[]string{"type"},
	)

	m.metrics.errorCount = prometheus.NewCounter(prometheus.CounterOpts{
		Name: "mempool_errors_total",
		Help: "Total number of errors",
	})

	m.metrics.queueDepth = prometheus.NewGauge(prometheus.GaugeOpts{
		Name: "mempool_queue_depth",
		Help: "Current depth of the transaction queue",
	})

	prometheus.MustRegister(m.metrics.transactionCount)
	prometheus.MustRegister(m.metrics.blockCount)
	prometheus.MustRegister(m.metrics.txLatency)
	prometheus.MustRegister(m.metrics.profitTotal)
	prometheus.MustRegister(m.metrics.profitPerTx)
	prometheus.MustRegister(m.metrics.memoryUsage)
	prometheus.MustRegister(m.metrics.goroutineCount)
	prometheus.MustRegister(m.metrics.opportunityCount)
	prometheus.MustRegister(m.metrics.errorCount)
	prometheus.MustRegister(m.metrics.queueDepth)
}

func (m *MempoolMonitor) collectSystemMetrics(ctx context.Context) {
	ticker := time.NewTicker(time.Minute)
	defer ticker.Stop()

	for {
		select {
		case <-ctx.Done():
			return
		case <-ticker.C:
			// Collect eBPF statistics
			slowSyscalls, avgLatency, maxLatency := m.ebpfMon.GetStats()
			
			// Collect network statistics
			rxPkts, txPkts, rxBytes, txBytes, rxDrops, txDrops, netLatency := m.netManager.GetStats()

			m.logger.Info("System metrics",
				// eBPF metrics
				zap.Uint64("slow_syscalls", slowSyscalls),
				zap.Duration("avg_syscall_latency", time.Duration(avgLatency)),
				zap.Duration("max_syscall_latency", time.Duration(maxLatency)),
				
				// Network metrics
				zap.Uint64("rx_packets", rxPkts),
				zap.Uint64("tx_packets", txPkts),
				zap.Uint64("rx_bytes", rxBytes),
				zap.Uint64("tx_bytes", txBytes),
				zap.Uint64("rx_drops", rxDrops),
				zap.Uint64("tx_drops", txDrops),
				zap.Duration("network_latency", time.Duration(netLatency)),
			)

			// Other metrics collection...
			m.metrics.memoryUsage.Set(float64(m.getMemoryUsage()))
			m.metrics.goroutineCount.Set(float64(runtime.NumGoroutine()))
		}
	}
}

func (m *MempoolMonitor) getMemoryUsage() uint64 {
	var memStats runtime.MemStats
	runtime.ReadMemStats(&memStats)
	return memStats.Alloc
}

func (m *MempoolMonitor) isRelevantTransaction(tx *Transaction) bool {
	if tx == nil {
		return false
	}

	// Check gas price
	if tx.GasPrice().Cmp(m.cfg.MaxGasPrice) > 0 {
		return false
	}

	// Check if transaction is already processed
	if m.cache.Contains(tx.Hash()) {
		return false
	}

	return true
}

func (m *MempoolMonitor) processingWorker(ctx context.Context) {
	for {
		select {
		case <-ctx.Done():
			return
		case tx := <-m.txChan:
			if err := m.processTransaction(tx); err != nil {
				m.logger.Error("Failed to process transaction", zap.Error(err))
				m.breaker.RecordError(err)
			}
		}
	}
}

type AffinityWorkerPool struct {
	workers int
	queue   chan func()
	wg      sync.WaitGroup
}

func NewAffinityWorkerPool(config *WorkerConfig) *AffinityWorkerPool {
	return &AffinityWorkerPool{
		workers: config.Workers,
		queue:   make(chan func(), config.QueueSize),
	}
}

func (p *AffinityWorkerPool) Start(ctx context.Context) error {
	for i := 0; i < p.workers; i++ {
		p.wg.Add(1)
		go p.worker(ctx)
	}
	return nil
}

func (p *AffinityWorkerPool) Stop() error {
	close(p.queue)
	p.wg.Wait()
	return nil
}

func (p *AffinityWorkerPool) Submit(task func()) {
	p.queue <- task
}

func (p *AffinityWorkerPool) worker(ctx context.Context) {
	defer p.wg.Done()

	for {
		select {
		case <-ctx.Done():
			return
		case task, ok := <-p.queue:
			if !ok {
				return
			}
			task()
		}
	}
}

type MMapPool struct {
	// ...
}

func NewMMapPool(path string) (*MMapPool, error) {
	// ...
}

func (m *MMapPool) Close() error {
	// ...
}

type CircuitBreaker struct {
	config      *config.CircuitBreakerConfig
	errorCount  atomic.Uint64
	lastReset   time.Time
	lastTripped time.Time
	enabled     atomic.Bool
	healthy     atomic.Bool
	mu          sync.RWMutex
	logger      *zap.Logger
	metrics     struct {
		tripCount    prometheus.Counter
		errorRate    prometheus.Counter
		healthyTime  prometheus.Gauge
		lastTripTime prometheus.Gauge
	}
}

func NewCircuitBreaker(cfg *config.CircuitBreakerConfig, logger *zap.Logger) *CircuitBreaker {
	cb := &CircuitBreaker{
		config:    cfg,
		logger:    logger,
		lastReset: time.Now(),
		enabled:   atomic.Bool{},
		healthy:   atomic.Bool{},
	}

	cb.metrics.tripCount = prometheus.NewCounter(prometheus.CounterOpts{
		Name: "circuit_breaker_trips_total",
		Help: "Total number of circuit breaker trips",
	})

	cb.metrics.errorRate = prometheus.NewCounter(prometheus.CounterOpts{
		Name: "circuit_breaker_errors_total",
		Help: "Total number of errors recorded by circuit breaker",
	})

	cb.metrics.healthyTime = prometheus.NewGauge(prometheus.GaugeOpts{
		Name: "circuit_breaker_healthy_time_seconds",
		Help: "Time since last healthy state in seconds",
	})

	cb.metrics.lastTripTime = prometheus.NewGauge(prometheus.GaugeOpts{
		Name: "circuit_breaker_last_trip_time_seconds",
		Help: "Time since last circuit breaker trip in seconds",
	})

	prometheus.MustRegister(cb.metrics.tripCount)
	prometheus.MustRegister(cb.metrics.errorRate)
	prometheus.MustRegister(cb.metrics.healthyTime)
	prometheus.MustRegister(cb.metrics.lastTripTime)

	cb.enabled.Store(true)
	cb.healthy.Store(true)

	go cb.monitor()

	return cb
}

func (cb *CircuitBreaker) RecordError(err error) bool {
	if !cb.enabled.Load() {
		return false
	}

	newCount := cb.errorCount.Add(1)

	// Check if we've exceeded the error threshold
	if int(newCount) >= cb.config.ErrorThreshold {
		cb.tripCircuit()
		return true
	}

	return false
}

func (cb *CircuitBreaker) tripCircuit() {
	cb.enabled.Store(false) // Set to false when tripped
	cb.healthy.Store(false)
	cb.lastTripped = time.Now()
	cb.metrics.tripCount.Inc()
	cb.metrics.lastTripTime.Set(time.Since(cb.lastTripped).Seconds())

	cb.logger.Warn("Circuit breaker tripped",
		zap.Uint64("error_count", cb.errorCount.Load()),
		zap.Duration("since_last_trip", time.Since(cb.lastTripped)))
}

func (cb *CircuitBreaker) monitor() {
	ticker := time.NewTicker(time.Second)
	defer ticker.Stop()

	for range ticker.C {
		cb.mu.Lock()

		// Check if we should reset
		if time.Since(cb.lastReset) >= cb.config.ResetInterval {
			cb.errorCount.Store(0)
			cb.lastReset = time.Now()

			// If we've been in cooldown for long enough, reset the circuit breaker
			if !cb.enabled.Load() && time.Since(cb.lastTripped) >= cb.config.CooldownPeriod {
				cb.enabled.Store(true) // Set to true when reset
				cb.healthy.Store(true)
				cb.logger.Info("Circuit breaker reset",
					zap.Duration("cooldown_period", cb.config.CooldownPeriod))
			}
		}

		// Update metrics
		if cb.healthy.Load() {
			cb.metrics.healthyTime.Set(time.Since(cb.lastTripped).Seconds())
		}

		cb.mu.Unlock()
	}
}

func (cb *CircuitBreaker) IsHealthy() bool {
	return cb.enabled.Load() && cb.healthy.Load()
}

func calculatePriority(tx *types.Transaction) float64 {
	// Basic priority calculation based on gas price
	// This will be enhanced with more sophisticated metrics
	return float64(tx.GasPrice().Uint64())
}

type ebpf struct {
	logger *zap.Logger
}

func NewSyscallMonitor(logger *zap.Logger) (*ebpf, error) {
	return &ebpf{logger: logger}, nil
}

func (e *ebpf) Start(ctx context.Context) error {
	// Start eBPF monitoring
	return nil
}

func (e *ebpf) Stop() error {
	// Stop eBPF monitoring
	return nil
}

func (e *ebpf) GetStats() (uint64, uint64, uint64) {
	// Get eBPF statistics
	return 0, 0, 0
}
